Adds test for adding person
