+++ test/test_amity.py
@@ -7,19 +7,6 @@ class TestAmity(TestCase):
     def setUp(self):
         self.room1 = Amity()
 
-    def test_room_wing(self):
-        '''
-        Ensures that only two type of living space wings are created,
-        that is the male_wing and female_wing
-        '''
-        pass
-
-    def test_add_person(self):
-        '''
-        Test that the person being added is not already in the system
-        '''
-        pass
-
     def test_load_people(self):
         '''
         Check that every person does not exist in the system before
